+++ generics/components/molecules/Molecule_ImgAndText.mjs
@@ -2,6 +2,7 @@ import { Component } from "../../../core/Component.mjs";
 import { slot } from "../../../core/helpers.mjs";
 import { Atom_Image } from "../atoms/Atom_Image.mjs";
 import { Atom_Text1 } from "../atoms/Atom_Text1.mjs";
+import { Modal_ProjectInfo } from "../organisms/Modal_ProjectInfo.mjs";
 
 export function Molecule_ImageAndText ( model )
 {
@@ -14,6 +15,7 @@ export function Molecule_ImageAndText ( model )
             <div class="molecule_image-and-text">
                     ${ slot( "img" ) }
                     ${ slot( "text1" ) }
+                <div id="modal-projectInfo"></div>
             </div>
         `;
     };
@@ -24,5 +26,18 @@ export function Molecule_ImageAndText ( model )
         let text1 = new Atom_Text1( model.atom_text1 );
         this.fillSlot( "img", img.getElement() );
         this.fillSlot( "text1", text1.getElement() );
+
+
+        this.getElement().querySelectorAll("molecule_image-and-text").addEventListener("click", (e) => {
+            console.log('btn-project button pressed')      
+            const modalProjInfo = document.getElementById('modal-projectInfo')
+            
+            modalProjInfo.innerHTML = `
+                ${slot("new-modal")}
+            `
+            this.modal = new Modal_ProjectInfo(model.content)
+
+            this.fillSlot("new-modal", this.modal.getElement());
+        });
     };
 }
